chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/getComputedStyle-with-distribution-expected.txt

crbug.com/337618: computedStyle should be cleared when distribution is updated

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL window.getComputedStyle(getNodeInComposedTree("child/bar")).color should be rgb(0, 128, 0). Was rgb(0, 0, 0).
PASS window.getComputedStyle(getNodeInComposedTree("child/bar")).color is "rgb(0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
